refactor: solve conflict
