+++ data/src/main/java/com/london/data/di/LocalDataSourceModule.kt
@@ -11,7 +11,7 @@ import com.london.data.local.database.dao.recent.whatched.tvshow.RecentWatchedTv
 import com.london.data.local.database.dao.search.SearchActorsDao
 import com.london.data.local.database.dao.search.SearchMoviesDao
 import com.london.data.local.database.dao.search.SearchTvShowDao
-import com.london.data.local.model.home.PopularSectionLocal
+import com.london.data.local.model.home.popular.PopularSectionLocal
 import com.london.data.local.model.home.TopRatedLocal
 import com.london.data.local.model.recent.search.RecentSearchLocal
 import com.london.data.local.model.recent.viewed.RecentViewedLocal
