+++ data/src/main/java/com/london/data/di/RepositoryModule.kt
@@ -1,7 +1,7 @@
 package com.london.data.di
 
 import com.london.data.local.database.dao.search.GenreInterestDao
-import com.london.data.local.model.home.PopularSectionLocal
+import com.london.data.local.model.home.popular.PopularSectionLocal
 import com.london.data.local.model.home.TopRatedLocal
 import com.london.data.local.model.recent.search.RecentSearchLocal
 import com.london.data.local.model.recent.viewed.RecentViewedLocal
