+++ data/src/main/java/com/london/data/local/database/dao/home/popular/PopularSectionDao.kt
@@ -4,7 +4,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.london.data.local.model.home.PopularSectionLocal
+import com.london.data.local.model.home.popular.PopularSectionLocal
 
 @Dao
 interface PopularSectionDao: HomeDao<PopularSectionLocal> {
