+++ data/src/main/java/com/london/data/repository/home/popular/PopularRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.london.data.repository.home.popular
 
-import com.london.data.local.model.home.PopularSectionLocal
+import com.london.data.local.model.home.popular.PopularSectionLocal
 import com.london.data.local.source.home.popular.HomeLocalDataSource
 import com.london.data.mapper.popular.toMovieEntity
 import com.london.data.mapper.popular.toPopularMovieSectionLocal
